fix: id

+++ src/main/java/ClassesNew/Cliente.java
@@ -4,7 +4,7 @@ public class Cliente extends Pessoa{
     
     private int ID_cliente;
     
-    public Cliente (int ID_pessoa, String nome, int telefone, String email, String cidade, String endereco, int ID_cliente) {
+    public Cliente (int ID_pessoa, String nome, long telefone, String email, String cidade, String endereco, int ID_cliente) {
         super(ID_pessoa, nome, telefone, email, cidade, endereco);
         this.ID_cliente = ID_cliente;
     }
